backend/requirements.txt
Django==4
psycopg2
djangorestframework
mip
numpy
django-cors-headers